consolidate black scholes and geo. brownian motion
Rename the black scholes process to geometric brownian motion, because
that's more recognizable. Remove the old geometric brownian motion,
which wasn't a technically correct implementation.
