+++ pyesg/__init__.py
@@ -3,7 +3,6 @@ from pyesg.interpolators.nelson_siegel import NelsonSiegelInterpolator
 from pyesg.interpolators.svensson import SvenssonInterpolator
 
 from pyesg.processes.academy_rate_process import AcademyRateProcess
-from pyesg.processes.black_scholes_process import BlackScholesProcess
 from pyesg.processes.cox_ingersoll_ross_process import CoxIngersollRossProcess
 from pyesg.processes.geometric_brownian_motion import GeometricBrownianMotion
 from pyesg.processes.heston_process import HestonProcess
@@ -15,7 +14,6 @@ __all__ = [
     "NelsonSiegelInterpolator",
     "SvenssonInterpolator",
     "AcademyRateProcess",
-    "BlackScholesProcess",
     "CoxIngersollRossProcess",
     "GeometricBrownianMotion",
     "HestonProcess",
